Ltac2 deep matches (unoptimized interpretation)
The system checks that the match is exhaustive printing an example
missing pattern when it's not (unlike ocaml it's an error not a
warning).

If the already existing code can handle the match we then intern it
that way. This keeps the same efficient evaluation for matches which
were already handled. Otherwise we have a naive interpreter where a
match is just a list of `pattern * expression` and we test each
pattern in order.

Not sure if it's worth keeping the optimized glb cases for simple
matches. I guess we should find some bench cases to know what the
difference looks like.

Empty matches only work if the type is empty in a shallow way, ie
`fun x => match x with end` works for `x:empty` but not for `x:empty * empty`.

Future work:
- useless pattern analysis (ie warn if the user write `None | None`)
- patterns for records, strings, ints, floats
- more efficient compilation